requirements.txt
appdirs==1.4.4
attrs==23.1.0
audioread==3.0.0
backcall==0.2.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
contextlib2==21.6.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.1.1
execnet==1.9.0
ffmpeg==1.4
ffmpeg-python==0.2.0
fonttools==4.38.0
future==0.18.3
idna==3.4
importlib-metadata==6.3.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.16.2
ipython==7.34.0
jams==0.3.4
jedi==0.18.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==4.12.0
keyboard==0.13.5
kiwisolver==1.4.4
lazy_loader==0.2
librosa==0.9.2
llvmlite==0.39.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mir-eval==0.7
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186163140/work
mkl-service==2.4.0
mock==5.0.2
msgpack==1.0.5
nest-asyncio==1.5.6
numba==0.56.4
numpy @ file:///C:/ci/numpy_and_numpy_base_1653574840943/work
packaging==23.0
pandas==1.3.5
parso==0.8.3
path==16.6.0
path.py==12.5.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
pluggy==1.2.0
pooch==1.6.0
prompt-toolkit==3.0.38
psutil==5.9.4
PyAudio==0.2.13
pycparser==2.21
pydub==0.25.1
Pygments==2.15.0
pynput==1.7.6
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.3.2
pytest-shutil==1.7.0
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
PyYAML==5.4.1
pyzmq==25.0.2
requests==2.28.2
resampy==0.4.2
scikit-learn==1.0.2
scipy @ file:///C:/ci/scipy_1661333074914/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sortedcontainers==2.4.0
soundata==0.1.2
soundfile==0.12.1
soxr==0.3.5
termcolor==2.3.0
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
wincertstore==0.2
zipp==3.15.0
